refactor: split useLocalBackup.ts into sub-hooks to resolve ESLint max-lines-per-function limit
